+++ src/routes/sis/courses/router.ts
@@ -1,10 +1,9 @@
 import { Router } from "express";
 
-import { UserRole } from "../../../generated/prisma";
+import { Prisma, UserRole } from "../../../generated/prisma";
 
-import { auth } from "../../../auth";
+import { auth, AuthRequest } from "../../../auth";
 import { courses } from "../../../sis";
-import { isNumber, isString } from "../../../utils";
 
 export const CourseRouter = Router();
 
@@ -13,15 +12,16 @@ CourseRouter.post(
     auth.authenticate,
     auth.authorization((user) => user.role == UserRole.Administrator),
     async (req, res) => {
-        const { code, name, units } = req.body;
-        if (!isString(code) || !isString(name) || !isNumber(units))
-            return res
-                .status(400)
-                .json({ message: "Invalid or missing course data" });
+        const { data } = req.body;
+        const result = courses.schema.safeParse(data);
 
+        if (!result.success)
+            return res.status(400).json({ message: "Invalid course data" });
+
+        const { code, name, units } = result.data;
         const course = await courses.create(code, name, units);
 
-        return res.status(course ? 200 : 403).json({
+        return res.status(course ? 200 : 409).json({
             message: course
                 ? "Course created successfully"
                 : "Course already exists",
@@ -30,7 +30,29 @@ CourseRouter.post(
     }
 );
 
-// TODO: CourseRouter.put()/CourseRouter.patch
+CourseRouter.put(
+    "/:code",
+    auth.authenticate,
+    auth.authorization((user) => user.role == UserRole.Administrator),
+    async (req: AuthRequest, res) => {
+        const { data } = req.body;
+        const result = courses.partialSchema.safeParse(data);
+
+        if (!result.success)
+            return res.status(400).json({ message: "Invalid course data" });
+
+        const success = await courses.update(
+            req.params.code as string,
+            result.data as Prisma.CourseUpdateInput
+        );
+
+        return res.status(success ? 200 : 404).json({
+            message: success
+                ? "Updated the course successfully"
+                : "Course does not exist",
+        });
+    }
+);
 
 CourseRouter.delete(
     "/:code",
@@ -38,13 +60,13 @@ CourseRouter.delete(
     auth.authorization((user) => user.role == UserRole.Administrator),
     async (req, res) => {
         const { code } = req.params;
-        if (!isString(code))
+        if (!code)
             return res
                 .status(400)
                 .json({ message: "Invalid or missing course code" });
 
         const course = await courses.delete(code);
-        return res.status(course ? 200 : 403).json({
+        return res.status(course ? 200 : 404).json({
             message: course
                 ? "Course deleted successfully"
                 : "Course does not exist",
@@ -59,7 +81,7 @@ CourseRouter.get(
     auth.authorization((user) => user.role == UserRole.Administrator),
     async (req, res) => {
         const { code } = req.params;
-        if (!isString(code))
+        if (!code)
             return res
                 .status(400)
                 .json({ message: "Invalid or missing course code" });