+++ src/routes/sis/sections/index.ts
@@ -1,0 +1,4 @@
+import { SISRouter } from "../router";
+import { SectionRouter } from "./router";
+
+SISRouter.use("/sections", SectionRouter);