+++ src/sis/index.ts
@@ -1,3 +1,4 @@
 export * from "./course";
 export * from "./program";
+export * from "./section";
 export * from "./student";